feat: macOS and bsd reuse the same code for there are just a little of differences, and add support for Android
